fix(inode): apply the epoch-extension bits and read timestamps as signed
G2 from docs/format-conformance-gaps.md. Two independent errors in the
same fields, both silent.

FIRST: the *_extra fields were read only for their nanosecond half.

  ctime_nsec = u32::from_le_bytes(raw[0x84..0x88]...) >> 2;

The format puts the SECONDS EXTENSION in the low two bits of that same
word -- exactly the bits `>> 2` discards. Those two bits widen the
seconds by 2^32 each, moving the ceiling from 2038 to roughly 2446. The
crate's own doc comment recorded the layout correctly ("top 30 bits =
nsec, low 2 bits = epoch") and then did not use it, so every timestamp
past 2038 came back 136 years early.

SECOND: the base field is SIGNED on disk, and was stored as u32. A date
before 1970 is a legal negative value; read unsigned it becomes a date
in 2106.

decode_extra_time applies both, matching ext4_decode_extra_time in
fs/ext4/ext4.h. Inode's timestamp fields widen u32 -> i64.

BREAKING, AND NOT ONLY FOR RUST. fs_ext4_attr_t's atime/mtime/ctime/
crtime widen to int64_t, which CHANGES THE C STRUCT LAYOUT -- every
field after them moves. include/fs_ext4.h is updated to match and says
so; consumers must recompile, not merely relink. Needs 0.6.0.

MY FIRST TEST WAS WRONG AND FAILING IT TAUGHT ME THE ENCODING. I
constructed 2100 by splitting the value at bit 32, giving epoch=0. The
kernel encodes it as

  extra = ((tv_sec - (s32)tv_sec) >> 32) & EXT4_EPOCH_MASK

so the epoch bits account for the signed reinterpretation of the base,
not for bits above 32. 2100 fits in a u32 but is negative as an i32, so
it is stored as that negative base PLUS an epoch of 1, and the two
cancel. The test now uses the kernel's formula and round-trips across
1901, 1969, 1970, 2000, the 2038 ceiling, one second past it, 2100, and
a value needing both epoch bits.

805 tests pass. Mutation-checked: discarding the epoch bits fails 3,
reading the base unsigned fails 3.

+++ include/fs_ext4.h
@@ -55,10 +55,19 @@ typedef struct {
     uint32_t uid;
     uint32_t gid;
     uint64_t size;
-    uint32_t atime;
-    uint32_t mtime;
-    uint32_t ctime;
-    uint32_t crtime;        /* Creation time (ext4 extra) */
+    /* Seconds since the Unix epoch, signed and 64-bit.
+     *
+     * Widened from uint32_t in 0.6.0. ext4's on-disk base field is a
+     * SIGNED 32-bit value which the matching *_extra field extends by
+     * two further bits, so the real range is roughly 1901..2446. A
+     * uint32_t truncated everything past 2038 and turned every
+     * pre-1970 date into a far-future one.
+     *
+     * THIS CHANGES THE STRUCT LAYOUT. Recompile any consumer. */
+    int64_t  atime;
+    int64_t  mtime;
+    int64_t  ctime;
+    int64_t  crtime;        /* Creation time (ext4 extra) */
     uint16_t link_count;
     fs_ext4_file_type_t file_type;
     /* Sub-second timestamp precision (added in v0.3) */
